More cleaning
And customizable collision height
