--- res/models/UFO.mtl
@@ -1,58 +1,0 @@
-# Blender MTL File: 'UFO2.blend'
-# Material Count: 6
-
-newmtl Material.001
-Ns 96.078431
-Ka 0.000000 0.000000 0.000000
-Kd 0.149646 0.177695 1.000000
-Ks 0.000000 0.000000 0.000000
-Ni 1.000000
-d 1.000000
-illum 0
-
-newmtl Material.002
-Ns 96.078431
-Ka 0.000000 0.000000 0.000000
-Kd 0.800000 0.000000 0.000000
-Ks 0.003268 0.003268 0.003268
-Ni 1.000000
-d 1.000000
-illum 0
-
-newmtl Material.003
-Ns 96.078431
-Ka 0.000000 0.000000 0.000000
-Kd 0.000000 0.983498 1.000000
-Ks 0.000000 0.000000 0.000000
-Ni 1.000000
-d 1.000000
-illum 0
-
-newmtl Material.005
-Ns 96.078431
-Ka 0.000000 0.000000 0.000000
-Kd 1.000000 1.000000 1.000000
-Ks 0.000000 0.000000 0.000000
-Ni 1.000000
-d 1.000000
-illum 0
-
-newmtl Material.006
-Ns 96.078431
-Ka 0.000000 0.000000 0.000000
-Kd 1.000000 1.000000 1.000000
-Ks 0.000000 0.000000 0.000000
-Ni 1.000000
-d 1.000000
-illum 0
-map_Kd 931377098.jpg
-
-newmtl Material.008
-Ns 96.078431
-Ka 0.000000 0.000000 0.000000
-Kd 0.800000 0.800000 0.800000
-Ks 0.450450 0.450450 0.450450
-Ni 1.000000
-d 1.000000
-illum 2
-map_Kd 931377098.jpg
