--- res/models/cube.obj
@@ -1,52 +1,0 @@
-v 1.000000 -1.000000 -1.000000
-v 1.000000 -1.000000 1.000000
-v -1.000000 -1.000000 1.000000
-v -1.000000 -1.000000 -1.000000
-v 1.000000 1.000000 -1.000000
-v 1.000000 1.000000 1.000000
-v -1.000000 1.000000 1.000000
-v -1.000000 1.000000 -1.000000
-v 1.000000 -1.000000 -1.000000
-v 1.000000 -1.000000 -1.000000
-v 1.000000 -1.000000 1.000000
-v 1.000000 -1.000000 1.000000
-v -1.000000 -1.000000 -1.000000
-v -1.000000 -1.000000 -1.000000
-v 1.000000 1.000000 -1.000000
-v 1.000000 1.000000 -1.000000
-v -1.000000 -1.000000 1.000000
-v -1.000000 -1.000000 1.000000
-v 1.000000 1.000000 1.000000
-v 1.000000 1.000000 1.000000
-v -1.000000 1.000000 1.000000
-v -1.000000 1.000000 1.000000
-v -1.000000 1.000000 -1.000000
-v -1.000000 1.000000 -1.000000
-vt 0.500000 1.000000
-vt 0.500000 0.500000
-vt 1.000000 0.500000
-vt 0.000000 0.500000
-vt 0.000000 0.000000
-vt 0.500000 0.000000
-vt 0.000000 1.000000
-vt 0.000200 1.000000
-vt 0.500000 1.000000
-vt 1.000000 1.000000
-vn 0.000000 -1.000000 0.000000
-vn 0.000000 1.000000 0.000000
-vn 1.000000 0.000000 0.000000
-vn 0.000000 0.000000 1.000000
-vn -1.000000 0.000000 0.000000
-vn 0.000000 0.000000 -1.000000
-f 11/1/1 17/2/1 13/3/1
-f 24/4/2 22/5/2 20/6/2
-f 15/1/3 19/7/3 12/4/3
-f 6/1/4 21/8/4 18/4/4
-f 3/2/5 7/9/5 23/7/5
-f 1/4/6 4/2/6 8/1/6
-f 9/10/1 11/1/1 13/3/1
-f 16/2/2 24/4/2 20/6/2
-f 10/2/3 15/1/3 12/4/3
-f 2/2/4 6/1/4 18/4/4
-f 14/4/5 3/2/5 23/7/5
-f 5/7/6 1/4/6 8/1/6
